+++ tilesets/tilesets.js
@@ -1,0 +1,9 @@
+{
+    tilesetNames = [
+        {folder: "default", name: 'Original'}
+        ,{folder: "lostgift", name: 'Lost Gifts'}
+        //,{folder: "new folder", name: 'new name'}
+    ]
+    
+    //["FOLDER NAME", "SHOWN NAME"]
+}